pyproject.toml
[project]
name = "gemini-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp (>=1.12.3,<2.0.0)",
    "google-genai (>=1.28.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "gemini_mcp", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
